Feature/checkout 2 (#48)
* add checkout

* add checkout

* save bills

* add test

* adjust findLastNumberDay

* fix bills entities and dtos

* adjust findLastNumberDay

* persist bill

* persist bill

* save payment and prduct

* persist bill

* persist bill

* fix testes

* fix testes

* validations

* validate if product is disponible

* validate if product is disponible

* code review

* code review

* code review image

---------

+++ app/application/src/main/kotlin/br/com/fiap/gerenciamentopedidos/application/requests/EditarProdutoRequest.kt
@@ -2,13 +2,14 @@ package br.com.fiap.gerenciamentopedidos.application.requests
 
 import br.com.fiap.gerenciamentopedidos.domain.enums.Categoria
 import br.com.fiap.gerenciamentopedidos.domain.models.Produto
+import java.math.BigDecimal
 
 data class EditarProdutoRequest(
     val id: Long?,
     val nome: String? = null,
     val descricao: String? = null,
     val categoria: Categoria? = null,
-    val valor: Double? = null,
+    val valor: BigDecimal? = null,
     val tempoPreparo: Long? = null,
     val imagem: EditarImagemRequest? = null,
 ) {
